fail_on(): move internal implementation to avocado.core.decorators
The `avocado.core.decorators` module has been recently added with the
`skip*` family of decorators.  Still, we have `fail_on()`, also a
decorator, defined in `avocado.core.exceptions`.

For consistency, let's put them in the same module.  There are no
changes to users of the Avocado Test API.
