+++ avocado/__init__.py
@@ -25,7 +25,7 @@ __all__ = ['main',
 from avocado.core.job import main
 from avocado.core.test import Test
 from avocado.core.version import VERSION
-from avocado.core.exceptions import fail_on
+from avocado.core.decorators import fail_on
 from avocado.core.decorators import skip
 from avocado.core.decorators import skipIf
 from avocado.core.decorators import skipUnless
